DZ/server/requirements.txt
Django==3.1.1
mysqlclient==2.0.1
django-mysql==3.8.1
djangorestframework==3.13.1
django-cors-headers==3.10.1
drf-writable-nested==0.6.3